+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for com.google.play.integrity
 
+## [1.4.0] - 2024-10-14
+- Updated Java Play Integrity API dependency from 1.3.0 to 1.4.0.
+- Add verdict opt-out to standard requests.
+- Add [remediation dialogs](https://developer.android.com/google/play/integrity/remediation).
+
 ## [1.3.3] - 2024-09-18
 - Updated Play Common dependency to 1.9.2.
 - Updated Play Core dependency to 1.8.5.